Fix placeholder thread number addition in ressource_classes (workaround)
The ressource_classes() needs the number of thread but it also needs to
add 1 to this number. The problem is that this value is still a
placeholder, so the addition can't be done at this time and place.

Dirty workaround: added an $expr($nthreads + 1), to be interpreted on
the command line, so after the substitution. Should work as long as the
interpreter (i.e. bash) can expand it.