Update AlignmentXref_conf.pm
Bumping the memory allocated to DumpProteome as the default (4Gb) fails for some plants. Not sure how severely this would negatively impact a Fungi run, for example. I figure my time spent debugging the pipeline in the case of MEMLIMIT is more valuable than 'hands off time', so I always air on the side of requesting more memory.